requirements/dev.txt
-r test.txt

tox
flake8
django-extensions
django-debug-toolbar

